Refactor source code folders
