+++ src/css/styles.css
@@ -10,5 +10,5 @@ body {
   position: absolute;
   top: 0; left: 0; right: 0; bottom: 0;
   margin: auto;
-  background: repeat url('../img/background.png');
+  background: repeat url('../assets/img/background.png');
 }
