+++ src/game/particles/explosion.ts
