+++ src/game/scenes/game/common/rotableBodySprite.ts
