+++ src/game/scenes/game/player/proyectileSprite.ts
